feat: custom logger lib created
A logging system is a must for this project, for this reason
one has been created and tested. It has some characteristics
that will improve by the time. For now it has only 4 levels
of loggin: debug, info, warn and error. In the future it'll
have more features. But for this project it's perfect as is.